first cut at refactoring to handle multiple sources in a db
when dealing with binary/source, the binary and source may be out of sync,
hence, need to track both sets of packages. For example,
even though devtools 2.0 is available as source, only
devtools 1.13 is availabe as osx binary in Nov 2018.

The current implementation only queries the source PACKAGES db, then looks for the same version as a binary
if the binary version is set. The objective here is
to refactor the entire internals to track the source and binary PACKAGES separately
so packages can be tracked properly, as well as control
specific versions, so in the above example could default to binary but
use the more recent source devtools during the install.