wip: rewriting functions for updating DESCRIPTION files. Not complete.
The functions work, but newlines are not preserved so the new DESCRIPTION file is just one-line. Some quick-ish research seems to indicate that Go does not have a reliable cross-platform solution to newline characters (most of the functions that should do this apparently still use \n in the background, at least as of 2018). I hoped to get around this by using bufio WriteString line-for-line, but somewhere along the way newlines are being dropped. I'll keep plugging away at this.