wip: first half of more simple approach to tarball dependency handling.
I was previously trying to maintain dependency constraints for package tarballs, but after realizing that pkgr isn't checking those, I decided
to simply add the dependencies as user packages like Devin and I discussed. I have also in this commit changed the automated tests
and added a localrepo for the case of a tarball that has dependencies.