change: changing from a worker concurrency model to a semaphore concurrency model.
also changing default thread count to the Number of CPUs (was previously zero). This was done to fix a bug where the program would try to launch with zero threads and would therefore hang. We need to verify that the default thread behavior for the 'threads' integration tests works as expected.