METWORX-202 Strip conda from the front of the path if present
Since R uses its own compiler/linker configuration, we have the
option of modifying Renviron and Makeconf, but this is not a 100%
reliable solution for mitigating problems.

Conda ships more or less as a complete chroot including system
utilities, and its own version of many libraries. We should not
let Conda override this, so strip it from the path instead. For
configurability, add an option to let users preserve this
(possible that it will never be used, but it's there)