exp: allow for ignoring certain packages in the solve
this will currently rip them out of the dep tree, so an ignored package
and all its deps will be removed. This means if that package is part of
the required package set it will cause the solve to fail, however
when having some suggested deps that want to selectively exclude
can work well.