refactor: use noupdate to normalize to the inverse pattern elsewhere and change default
it has found that the default behavior to not update is generally not what users
want, especially around reproducibility. Instead, we want users to be
able to not update only if they absolutely need to.

In addition, part of this original decision was to prevent packages attempting to be
updated causing major failures that couldn't get back to original state.
This is no longer an issue given the rollback functionality.