config: invoke renv to discover library path
As of v0.15 [1], renv's default behavior is to put the library of a
_package_ project outside of the main project directory.  On the other
hand, pkgr, when fed a `Lockfile: Type: renv` configuration, installs
to the old renv/library/... location.  Users that do not change renv's
default will not see the packages installed by pkgr.

Aside from stating that pkgr won't support renv's new default behavior
and telling users to point `RENV_PATHS_LIBRARY_ROOT` back to the old
default, the two main options to fix this are

  * teach pkgr to mimic renv's logic
  * invoke renv to get the library path from it

renv's logic is non-trivial, including things like embedding the hash
of the full project path in the external project directory name,
checking whether a renv profile is active (via either the environment
variable `RENV_PROFILE` or the file renv/profile), checking whether
`RENV_PATHS_LIBRARY_ROOT` is set, and calling out to
`tools$R_user_dir` to get a system-specific behavior.

pkgr could repeat that logic, accounting for renv's different behavior
before and after v0.15.  But there's probably a good chance that, even
if the logic is ported correctly, it will get out of date with renv's
logic.

To avoid that issue, this commit goes with the second approach, asking
renv for the path.  It does so through the `renv::paths` API that has
been available since since renv v0.7.0 (2019).  If the call fails,
pkgr falls back to the previous behavior.

The main downside of this approach is that it slows down pkgr.  The
worst case scenario is renv needing to bootstrap itself (which can
take several seconds).  That's not great, but it's a one-off cost that
the renv use would pay anyway the next time they start an R session.

However, even for the non-bootstrap case, invoking R to get
information from renv unsurprisingly slows things down:

  $ hyperfine -n pr -n v3.0.0-rc1 --warmup 2 \
    '~/go/bin/pkgr plan' '/usr/local/bin/pkgr plan'
  Benchmark 1: pr
    Time (mean ± σ):     306.6 ms ±   5.5 ms    [User: 242.3 ms, System: 67.0 ms]
    Range (min … max):   301.4 ms … 321.0 ms    10 runs

  Benchmark 2: v3.0.0-rc1
    Time (mean ± σ):      44.9 ms ±   2.4 ms    [User: 41.8 ms, System: 13.4 ms]
    Range (min … max):    40.5 ms …  50.9 ms    57 runs

  Summary
    'v3.0.0-rc1' ran
      6.82 ± 0.38 times faster than 'pr'

While that difference will usually get swamped in the `pkgr install`
context, it is quite noticeable when calling `pkgr plan`.  Note,
though, that this cost is paid only for the `Lockfile: Type: renv`
code path.

Signal an error when the renv call fails [2]; this catches the case
when the system renv isn't available _and_ the project does not track
.Rprofile and renv/activate.R.  Another option would be to install to
the old $project/renv/library location, but that's unlikely to be the
correct location, especially over time as the renv <0.15 installs
become less common.

See gh-396 for more discussion on the decision to error here and on
different system renv + project renv scenarios.

Fixes #392.

[1] specifically 2de06e37 (use external project library for package
    projects, 2021-12-08)

[2] Ideally getLibraryPath() would return an error rather than calling
    log.Fatal(), moving the Fatal() call higher up the
    getLibraryPath() <- NewConfig() <- initConfig() call chain.
    However, NewConfig() already calls another function that dies
    rather than returning an error (loadConfigFromPath), so calling
    Fatal() in getLibraryPath() doesn't change much aside from making
    the failure case hard to test in the TestGetLibraryPath unit test.