TestGetLibraryPath: loosen renv library path expectation
An upcoming commit will teach getLibraryPath() to invoke renv to get
the library path.  Rather than updating the test to match renv's
system naming scheme just check that the path contains
"renv/library/R-X.Y/".  That should be good enough for the purpose of
this test (and there will be an integration test that checks that the
pkgr-installed renv library is functional).