config_test.go: don't assume top-level directory is under pkgr/
getTestFolder() and getIntegrationTestFolder() find the top-level
directory of the pkgr sources by splitting

    /full/path/pkgr/configlib/config_test.go

after "/pkgr/".

This fails if the top-level directory isn't named "pkgr", as is the
case on drone, where the sources are checked out to "/drone/src/" [*].

Instead get the top-level directory by constructing a relative path
from the full path to configlib/config_test.go.

[*] https://github-drone.metrumrg.com/metrumresearchgroup/pkgr/217/1/4