config_test.go: don't force renv case onto existing tests
TestNewConfigPackrat and TestGetLibraryPath are table-driven tests
that include renv as one of the cases.  With f5c120d0 (config: invoke
renv to discover library path, 2022-01-21), the logic for the renv
case is more involved than the other cases.

As pointed out by @Dreznel, doing special handling for renv as part of
the existing tests isn't a good use of table-driven tests, and the
downstream branching is difficult to follow.

Instead move the renv bits to dedicated tests, and restore
TestNewConfigPackrat and TestGetLibraryPath to (the equivalent of)
what they were before this series.

Re: https://github.com/metrumresearchgroup/pkgr/pull/396#discussion_r828088067