NEWS.md: copy over GitHub release notes for v3.0.0
After the v3.0.0 release, @seth127 filled in the GitHub release with a
user-facing description of the changes.  That was the first pkgr
release with release notes.

The plan is to add release notes going forward.  Given we're going to
be taking the time to write them, let's also track them in the repo.
This has the benefit of 1) allowing them to be reviewed and tweaked
before the release is cut and 2) lets people the have a checkout of
the repo see them without visiting GitHub.