config: mark renv library path for more reliable detection
As of f5c120d (config: invoke renv to discover library path,
2022-01-21), getRenvLibrary() finds the project library path by
invoking an renv function and writing the result to stdout.
getRenvLibrary() aborts if there's more than one line of output, so
this approach fails if any of the startup files write to stdout.

To handle this case, don't restrict stdout to a single line.  Instead
surround the renv function's output with markers that are unlikely to
be emitted by startup code.