tests: replace some Microsoft CRAN URLs to MPN URLs
Unlike the previous commits, these changes are not required because
these tests do not actually reach out to the site.  But, an MPN URL
works just as well here and avoids later confusion about whether the
dead URL is an issue.

After this, cran.microsoft.com URLs remain in a few spots:

 * integration_tests subdirectories of pkgr.yml files that do _not_
   have a Makefile wired up to run tests (unlike the tests run via the
   Drone)

 * a snippet of output in integration_tests/multi-repo/guide.md

 * the PkgrRepositoryURL field in the DESCRIPTION file of tracked
   libraries

 * a repository in localrepos/pkgr-retrieve.yml

These don't affect the automated tests, so let's leave these be for
now.  It's not clear how any of these should be updated, if at all
(and many of these files should probably just be removed from the
repo).