add target to check which Go files missing from coverage JSON
The scripts for VT_TEST_RUNNERS need to implement fairly complex
handling.  When GOCOVERDIR is set, they need to instrument a binary
for integration tests and to pass '-args -test.gocoverdir=$d' for unit
tests.  They need to do custom package selection, separating unit
tests from integration tests and removing this valtools subtree.  And
for unit tests, they need to pass -coverpkgs in order for the tests of
one package to count toward the coverage of another package.

On top of that, the files of packages without tests are not included
in the coverage profiles for unit tests invoked with -test.gocoverdir.
Take the following module:

    |-- bar
    |   `-- bar.go
    |-- foo
    |   |-- foo.go
    |   `-- foo_test.go
    `-- go.mod

If 'go test -coverprofile coverage.out ./...' is called from the top
level, the profile will include entries for bar.go.  However, if
'go test -cover ./... -args -test.gocoverdir=$d' followed by
'go tool covdata textfmt -i $d -o coverage.out' is called to create
the text format profile, entries for bar.go are _not_ present [*].

Given the above complexity, it's helpful to have a way to verify that
the Go files in the final coverage JSON match what's expected.

[*] tested under Go 1.22.5, which includes d1cb5c0605 (cmd/go: improve
    handling of no-test packages for coverage, 2023-05-09)