integration_tests/library: shorten a subtest name
If renv is available (as it is on Metworx), we skip the
TestLibraryRenv subtest that checks an error when renv is unavailable.
This is the one subtest name that makes it through to the scorecard
because it's skipped.

Reword the subtest to make it a bit shorter and remove the test ID.
(In general, we should go through the test suite in remove the test
IDs, which are no longer relevant under the current validation
approach, but this was the only test ID that made it to the
scorecard.)