docs: define a traceability matrix
This includes all the test files listed in the previous validation
docs except for integration_tests/baseline/root_test.go.  That tested
the --version option (which is covered by integration_tests/version).
If we want to list the test in the traceability matrix, we could
include it under the top-level pkgr entry (which would seem a bit odd
given that's the entry point to everything) or introduce a version
subcommand.  But leave it be for now.