ci: switch to GitHub Actions
With the go.mod update in this series, building pkgr on Drone is now
failing with several error messages that say

  //go:build comment without // +build comment

Based on a quick search, it seems like a bug with Go 1.16 (see
golang's issue 51436).  That lines up with the Go version we pull in
to build the binary and the Go version bundled with the r-dev-ci
images.

We were planning to switch to GitHub Actions soon, but, with the above
Drone error, it makes sense to do with this series.

The Drone build primarily tested on r-dev-ci:4.1.0.  The exceptions
were

 * r-dev-ci:4.0.5 for the mixed-source tests

 * r-dev-ci-mpn-4.1:2022-02-11 and r-dev-ci-mpn-4.1:cran-latest for
   the renv library checks

For the first point, using 4.0.5 is no longer necessary now that
TestMixedSource generates its own repo.

Under GitHub Actions, handle the 'renv vs no renv' by setting up two
different jobs, one with a system-wide renv installation and one
without.  Unlike Drone, this leaves out testing an renv version before
0.15 [*], which should be okay given the renv 0.15 release is now over
two years old.

Note that the library for renv is set up via Rprofile.site rather than
communicated through an environment variable (e.g., R_LIBS_USER or
R_LIBS_SITE) because getRenvLibrary uses RunRscriptWithArgs, which
intentionally discards these.  (getRenvLibrary should probably be
adjusted to avoid that behavior.)

For the OS, hold off on adding a job for the latest Ubuntu because
that triggers several 'pkgr install' tests to fail due to a fansi
build error.

[*] See 0a4a8ee (ci: test configlib and integration_tests/library with
    system renv, 2022-03-08) and f5c120d (config: invoke renv to
    discover library path, 2022-01-21) for why testing an renv version
    before 0.15 was of interest.