packrat: delete package
The last import of the packrat package was removed when the top-level
main.go was deleted in 8e2466a (chore: module and main cleanup,
2019-05-25).  Although there is an old issue [*] for adding the
ability to "parse packrat lock files and incorporate them into an
InstallPlan", I'm not aware of any active plans to work on that, and I
think that's unlikely to happen given RStudio says packrat "has been
soft-deprecated and is now superseded by renv".

And we can of course resurrect this code if needed (and this commit
should be easy to find, as it will be auto-linked to issue 87).

Note that this package is not related to the "Lockfile: Type: packrat"
feature for installing to a packrat-like location.  That's still
supported and tested in integration_tests/library.

[*] https://github.com/metrumresearchgroup/pkgr/issues/87