inspect: hide --installed-from flag
When --installed-from is passed, inspect outputs information about
whether the package was installed by pkgr or something else.  However,
the implementation isn't fully fleshed out, and the output is
confusing.

It outputs a result with the form

  {
    "pkgr": [...],
    "packrat": [...],
    "unknown": [...]
  }

The "pkgr" list is accurate and useful.  On the other hand, "unknown"
is always empty, and the "packrat" list includes packages that weren't
installed with packrat (e.g., a package installed directly with 'R CMD
INSTALL').