site: reformat command options as bullets
Cobra's GenMarkdownTree puts options in a code block, with each option
on a line. Options with anything but very short descriptions can flow
off the screen.

Displaying options as bullets, with the description as regular text,
would be more readable. One option would be to write a custom variant
of GenMarkdownTree, but, at least for now, just post-process the
markdown with a custom script.

Note that the display of long descriptions could also be improved for
the command-line help, but that'd involve an independent change
(registering a custom usage template via Cobra's SetUsageTemplate).