install_test.go: use pillar instead of testthat as tarball input
TestTarballInstall has a test case that confirms that a library is
correctly set up when there's a tarball specified for a package that
has dependencies.  It uses testthat 2.1.1 as the input tarball.

Compiling testthat 2.1.1 on Ubuntu 22.04 fails with

  ../inst/include/testthat/vendor/catch.h:6490:33: error: size of array ‘altStackMem’ is not an integral constant-expression
   6490 |         static char altStackMem[SIGSTKSZ];
        |                                 ^~~~~~~~
  ../inst/include/testthat/vendor/catch.h:6541:45: error: size of array ‘altStackMem’ is not an integral constant-expression
   6541 |     char FatalConditionHandler::altStackMem[SIGSTKSZ] = {};
        |                                             ^~~~~~~~

testthat 3.1.0 fixes that failure, so we could update the
localrepos/tarball-deps repo to use testthat 3.1.0 or later.  The
downside of that is that testthat 3.1.0 specifies 21 packages for its
Imports versus the 9 specified by testthat 2.1.1.  In addition to
further bloating the Git repository (our tests would ideally be going
in the other direction), adding more packages increases the odds of
hitting into incompatibilities with OS and R version updates.

Instead rearrange things to work with the source tarballs that we
already track while maintaining the core features of the test.  In
particular, the test should confirm that dependencies of the specified
tarball are installed to the library.  There is also a package
specified in the normal "Packages" set.  The dependencies of the
Packages package should not include all the dependencies of the
Tarballs package because then it would not be clear that the tarball
led to the dependency being installed in the library.

Use pillar (which has several dependencies) as the Tarballs package
and openssl as the Packages package (which pulls in askpass and
indirectly sys, neither of which are pulled in by pillar).  Drop now
unnecessary packages from the repo and regenerate PACKAGES* with
tools::write_PACKAGES (under R 4.4).