tarballOps: move away from deprecated archiver package
The archiver package is no longer developed and is superseded by the
archives package.  archiver has an unaddressed CVE filed against it [*].
The tarball input can be constructed in ways that lead to extracted
files being created outside the intended destination directory.

In the context of pkgr, this path traversal vulnerability isn't high
risk.  The functionality involves users specifying local tarballs of R
packages to install, so the user is already trusting the tarball input
and the code within it.  However, the package being no longer
maintained is reason alone to find another approach.

The new archives package doesn't have the functionality provided by
archiver.Unarchive, so we need to implement the extraction logic
ourselves, similar to what we did before e691c443 (fix: changing to
mholt/arhiver for unarchiving tarballs [...], 2020-04-28).
Importantly the new logic should

 * avoid a path traversal vulnerability

 * not introduce a regression for the bug that was fixed by the switch
   to archiver in e691c443 (gh-259).

The advantage of using the archives package is that it takes care of
decompressing the various tar.* formats that 'R CMD build' supports
(gz, bz2, xz, and zst).

[*] CVE-2024-0406
    mholt/archiver issue 404
    https://github.com/metrumresearchgroup/pkgr/security/dependabot/11