integration_tests/mixed-source: use better names for repos
pkgr-linux.yml.template includes two repos, MPNJuly2020 and
MPNJune2021.  The key distinction between these repos is that one is
configured with "binary" as its type and the other with "source".

Rename the repos to highlight this distinction and to avoid needing to
update the name of the repo each time the date is updated (as it will
be in the next commit).

(Update the golden files by passing -update to 'go test'.)