integration_tests/baseline: swap package for indirect dependency check
A test in install_test.go checks that fansi is installed.  fansi is a
dependency of pillar, and the point is to check a case where a package
is not explicitly listed in the pkgr configuration but is pulled in as
a dependency.

pillar 1.10.0 dropped fansi as a dependency.  Instead use cli for this
check, which will continue to work when the next commit updates the
test to use a newer MPN snapshot.